Add Nucleus.Dashboard project and enhance logging
Introduces the Nucleus.Dashboard project with middleware, configuration, and sample endpoints. Updates Nucleus.Core to use Z.Dapper.Plus for bulk inserts, adds logging to request flush and log services, and improves database schema with an index for faster queries. Integrates dashboard middleware into TESTaPP and updates configuration for more flexible flush intervals.

+++ Nucleus.Core/Config/NucleusOptions.cs
@@ -26,7 +26,7 @@ public class NucleusOptions
     /// <summary>
     /// How often, in seconds, the background flush service writes accumulated logs to the database.
     /// </summary>
-    public int BatchFlushIntervalSeconds { get; set; } = 1;
+    public float BatchFlushIntervalSeconds { get; set; } = 1;
     
     /// <summary>
     /// Optional custom table name for request logs.
